fix: extract shared portfolio-core crate to dedupe snapshot logic (#615)
build_portfolio_snapshot and convert_to_base were fully duplicated between
src-tauri and portfolio-mcp, with the #611 fix for #609 already having to be
applied by hand in both copies. Extract both into a new portfolio-core
workspace crate (canonical types + snapshot/fx logic) that both crates
depend on and re-export, so there's exactly one implementation to fix going
forward.

The canonical implementation uses src-tauri's more-correct logic in two
places where the two copies had already drifted: the #609/#618 zero-price
handling, and the same-day-purchase cost-basis proxy for daily PnL. This
fixes portfolio-mcp's get_portfolio_snapshot to match the desktop app's
get_portfolio for the same data, closing the exact drift #615 warned about.

Frontend TS bindings (frontend/types/bindings) are byte-identical after the
refactor — the wire format is unchanged.
