+++ Cargo.toml
@@ -2,5 +2,6 @@
 members = [
     "src-tauri",
     "portfolio-mcp",
+    "portfolio-core",
 ]
 resolver = "2"
