+++ portfolio-mcp/src/types.rs
@@ -1,17 +1,14 @@
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
+// Canonical portfolio types shared with `src-tauri` — see `portfolio-core`
+// and #615. Re-exported here so every other module in this crate can keep
+// using `crate::types::X` unchanged.
+pub use portfolio_core::types::{
+    AccountType, AssetType, FxRate, Holding, HoldingId, PortfolioSnapshot, PriceData,
+};
+
 // ── ID newtypes ───────────────────────────────────────────────────────────────
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
-#[serde(transparent)]
-pub struct HoldingId(pub String);
-
-impl std::fmt::Display for HoldingId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.0.fmt(f)
-    }
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
 #[serde(transparent)]
@@ -65,83 +62,6 @@ impl std::str::FromStr for TransactionType {
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "lowercase")]
-pub enum AssetType {
-    Stock,
-    Etf,
-    Crypto,
-    Cash,
-}
-
-impl AssetType {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            AssetType::Stock => "stock",
-            AssetType::Etf => "etf",
-            AssetType::Crypto => "crypto",
-            AssetType::Cash => "cash",
-        }
-    }
-}
-
-impl std::str::FromStr for AssetType {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "stock" => Ok(AssetType::Stock),
-            "etf" => Ok(AssetType::Etf),
-            "crypto" => Ok(AssetType::Crypto),
-            "cash" => Ok(AssetType::Cash),
-            other => Err(format!("Unknown asset type: {other}")),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "lowercase")]
-pub enum AccountType {
-    Tfsa,
-    Rrsp,
-    Fhsa,
-    Taxable,
-    Crypto,
-    Cash,
-    Other,
-}
-
-impl AccountType {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            AccountType::Tfsa => "tfsa",
-            AccountType::Rrsp => "rrsp",
-            AccountType::Fhsa => "fhsa",
-            AccountType::Taxable => "taxable",
-            AccountType::Crypto => "crypto",
-            AccountType::Cash => "cash",
-            AccountType::Other => "other",
-        }
-    }
-}
-
-impl std::str::FromStr for AccountType {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "tfsa" => Ok(AccountType::Tfsa),
-            "rrsp" => Ok(AccountType::Rrsp),
-            "fhsa" => Ok(AccountType::Fhsa),
-            "taxable" => Ok(AccountType::Taxable),
-            "crypto" => Ok(AccountType::Crypto),
-            "cash" => Ok(AccountType::Cash),
-            "other" => Ok(AccountType::Other),
-            other => Err(format!("Unknown account type: {other}")),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "lowercase")]
 pub enum AlertDirection {
     Above,
     Below,
@@ -169,29 +89,6 @@ impl std::str::FromStr for AlertDirection {
 }
 
 // ── Core data types ───────────────────────────────────────────────────────────
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Holding {
-    pub id: HoldingId,
-    pub symbol: String,
-    pub name: String,
-    pub asset_type: AssetType,
-    pub account: AccountType,
-    pub account_id: Option<String>,
-    pub account_name: Option<String>,
-    pub quantity: f64,
-    pub cost_basis: f64,
-    pub currency: String,
-    pub exchange: String,
-    pub target_weight: Option<f64>,
-    pub created_at: String,
-    pub updated_at: String,
-    pub indicated_annual_dividend: Option<f64>,
-    pub indicated_annual_dividend_currency: Option<String>,
-    pub dividend_frequency: Option<String>,
-    pub maturity_date: Option<String>,
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -255,83 +152,6 @@ pub struct PriceAlertInput {
     pub threshold: f64,
     pub currency: String,
     pub note: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PriceData {
-    pub symbol: String,
-    pub price: f64,
-    pub currency: String,
-    pub change: f64,
-    pub change_percent: f64,
-    pub updated_at: String,
-    pub open: Option<f64>,
-    pub previous_close: Option<f64>,
-    pub volume: Option<i64>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FxRate {
-    pub pair: String,
-    pub rate: f64,
-    pub updated_at: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct HoldingWithPrice {
-    pub id: HoldingId,
-    pub symbol: String,
-    pub name: String,
-    pub asset_type: AssetType,
-    pub account: AccountType,
-    pub account_id: Option<String>,
-    pub account_name: Option<String>,
-    pub quantity: f64,
-    pub cost_basis: f64,
-    pub currency: String,
-    pub exchange: String,
-    pub target_weight: Option<f64>,
-    pub created_at: String,
-    pub updated_at: String,
-    pub indicated_annual_dividend: Option<f64>,
-    pub indicated_annual_dividend_currency: Option<String>,
-    pub dividend_frequency: Option<String>,
-    pub maturity_date: Option<String>,
-    pub current_price: f64,
-    pub current_price_cad: f64,
-    pub market_value_cad: f64,
-    pub cost_value_cad: f64,
-    pub gain_loss: f64,
-    pub gain_loss_percent: f64,
-    pub weight: f64,
-    pub target_value: f64,
-    pub target_delta_value: f64,
-    pub target_delta_percent: f64,
-    pub daily_change_percent: f64,
-    pub fx_stale: bool,
-    pub price_is_stale: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PortfolioSnapshot {
-    pub holdings: Vec<HoldingWithPrice>,
-    pub total_value: f64,
-    pub total_cost: f64,
-    pub total_gain_loss: f64,
-    pub total_gain_loss_percent: f64,
-    pub daily_pnl: f64,
-    pub last_updated: String,
-    pub base_currency: String,
-    pub total_target_weight: f64,
-    pub target_cash_delta: f64,
-    pub realized_gains: f64,
-    pub annual_dividend_income: f64,
-    #[serde(default)]
-    pub requires_cost_basis_selection: bool,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
